move setup experience related integration tests to new file
